requirements.txt
azure-identity
azure-ai-ml
azure-storage-file-datalake
pyarrow
pandas
azure-cli
